feat: use ruamel.yaml
