+++ examples/dictknife/cat/src/00output-format/person.yaml
